Update the code style and rename a code file
