+++ src/test/ACCESS_CARTA_KNOWN_SESSION.test.ts
@@ -2,7 +2,7 @@ import { action, makeObservable, observable, runInAction } from 'mobx';
 import { CARTA } from 'carta-protobuf';
 import config from './config.json';
 
-const WebSocket = require('ws');
+import WebSocket from 'ws';
 let testServerUrl = config.serverURL0;
 let connectTimeout = config.timeout.connection;
 const icdVersion = config.icdVersion;
