+++ src/test/ACCESS_WEBSOCKET.test.ts
@@ -1,5 +1,5 @@
 import config from './config.json';
-const WebSocket = require('ws');
+import WebSocket from 'ws';
 
 let testServerUrl = config.serverURL0;
 let connectTimeout = config.timeout.connection;
